Refactored code to follow style and naming conventions
