--- Assets/Scripts/Line drawer/LineSettings.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 08f76b92fe3b91b4fbfb3767e06f98cd
