+++ Assets/Scripts/LineDrawer/LineSettings.cs.meta
@@ -1,0 +1,15 @@
+fileFormatVersion: 2
+guid: 08f76b92fe3b91b4fbfb3767e06f98cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences:
+  - functionLine: {instanceID: 0}
+  - trajectoryLine: {instanceID: 0}
+  - functionLineSettings: {fileID: 11400000, guid: b4f671f75320c974096cd68b849f36c3, type: 2}
+  - trajectoryLineSettings: {fileID: 11400000, guid: 1660e49522ad7044fa99866e5b8deb7b, type: 2}
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
