+++ Assets/Scripts/LineDrawer/ScriptableObjects/FunctionLineSettingVarables.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4f671f75320c974096cd68b849f36c3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
