+++ Assets/Scripts/Shake/Editor/ShakeBehaviorEditor.cs
@@ -1,19 +1,21 @@
 using UnityEngine;
 using UnityEditor;
-
-[CustomEditor(typeof(ShakeBehavior))]
-public class ShakeBehaviorEditor : Editor
+namespace Shake
 {
-    public override void OnInspectorGUI()
+    [CustomEditor(typeof(ShakeBehavior))]
+    public class ShakeBehaviorEditor : Editor
     {
-        base.OnInspectorGUI();
-
-        ShakeBehavior shakeScript = (ShakeBehavior)target;
-
-        if (GUILayout.Button("Test Shake"))
+        public override void OnInspectorGUI()
         {
-            // Trigger the shake using the current settings in the inspector
-            shakeScript.Shake(shakeScript.ShakeDuration, shakeScript.ShakeIntensity);
+            base.OnInspectorGUI();
+
+            var shakeScript = (ShakeBehavior)target;
+
+            if (GUILayout.Button("Test Shake"))
+            {
+                // Trigger the shake using the current settings in the inspector
+                shakeScript.Shake(shakeScript.ShakeDuration, shakeScript.ShakeIntensity);
+            }
         }
     }
 }
